refactor: rename ai remediation proposal worker
