+++ pkg/airemediationproposal/handler.go
@@ -1,4 +1,4 @@
-package airemediation
+package airemediationproposal
 
 import (
 	"context"
@@ -21,11 +21,11 @@ func NewSqsHandler(l logging.Logger) *SqsHandler {
 
 func (s *SqsHandler) HandleMessage(ctx context.Context, sqsMsg *types.Message) error {
 	msgBody := aws.ToString(sqsMsg.Body)
-	s.logger.Info(ctx, "got AI remediation message")
+	s.logger.Info(ctx, "got AI remediation proposal message")
 
 	msg, err := ParseQueueMessage(msgBody)
 	if err != nil {
-		s.logger.Errorf(ctx, "Invalid AI remediation message: SQS_msg=%+v, err=%+v", sqsMsg, err)
+		s.logger.Errorf(ctx, "Invalid AI remediation proposal message: SQS_msg=%+v, err=%+v", sqsMsg, err)
 		return mimosasqs.WrapNonRetryable(err)
 	}
 	if !common.IsMatchAccountIDArn(msg.AccountID, msg.AssumeRoleArn) {
@@ -41,7 +41,7 @@ func (s *SqsHandler) HandleMessage(ctx context.Context, sqsMsg *types.Message) e
 	}
 
 	s.logger.Infof(ctx,
-		"AI remediation worker skeleton received message, RequestID=%s, remediation_proposal_id=%d, finding_id=%d, project_id=%d, data_source=%s, aws_id=%d, aws_data_source_id=%d, account_id=%s",
+		"AI remediation proposal worker skeleton received message, RequestID=%s, remediation_proposal_id=%d, finding_id=%d, project_id=%d, data_source=%s, aws_id=%d, aws_data_source_id=%d, account_id=%s",
 		requestID,
 		msg.RemediationProposalID,
 		msg.FindingID,
